Use the new components for cultivation page
